[feat] add -disable-autostart flag and config
Make possible to disable the autostart configuration for the next run. I
can be done by configuring "DisableAustostart" to true in the
systra.json or by passing the '-disable-autostart' flag to the binary.

To make that possible I have rework how the Config struct work
separating the file configuration with the runtime configuration.

- Resolves: #100