[bug] exit cleanly in osx
two things happen differently in osx:

- call to systray.Quit() halts the program (so if called directly, none
        of the deferred functions that we use for cleanup get to
        execute)
- systray.Run() blocks (so after loop returns, the main run.Run() function
        did not get to receive the boolean through the finishedCh channel.

proper shutdown is therefore fixed here by moving the call to
systray.Quit() to a goroutine that executes when the initialize()
function calls all the deferred functions.

we need to revisit this in case we want to break the main select loop
for a reson other than successfully terminating the program.