[bug] accept corrupted pid files
A corrupted pid file, or a file that contains something else than a pid
number used to block bitmask-vpn from launch. Now if the file contains
something that is not expected it will log it and rewrite it.

I refactor de code so the pid file will be released on quit instead of
initialization.

- Resolves: #349