[feat] let's use the notification only in windows for now
We really need this for windows.

However, I'm afraid for regular users in linux/mac the notification
might be too obstrusive. As we commented in the mr, one way around this
might be to only display that tip the first few times - and then persist
a flag about it having been shown.