[pkg] abort install if we running an old version
stop if:

- bitmask is running (we don't want to mess with the helper if the app
  is running)
- we cannot stop the service for some reason.

powershell gives more flexibility to stop services, but Remove-Service
is not present in PS < 6, so we try to remove the service from the
post-install too (the golang helper will complain about an eventlog
registry key).

this could probably be improved in the helper, but we'll be moving to
the official openvpnserv2 service for 0.21.4. we might want to revisit
the current helper for the firewall/killswitch.