[bug] fail gracefully on connection initialization errors
There's some corner cases that were very badly captured. On the first
place, it's confusing to quit on connection errors.

Secondly, a side-effect of aborting the initialization of the bitmask
object was a semi-random segfault when trying to access the object.

Here I pass any connection errors to the gui, but leave to the user to
quit the app. This probably will need more work when we want to
terminate the app on unrecoverable errors (no polkit, etc...), but for
now it makes the systray much more usable.

- Resolves: #465