[feat] remove static ip as fallback mechanism #768
This commit removed the fallback codepath with the static ip. Discussion in #768

The hardcoded ip was also used for the Snowflake path. I removed it from there and use the DNS hostname. Therefore, the certificate pinning is gone and the SystemPool is used as default. I tested the snowflake setup and it worked.