Don't return gateway twice if location is selected manually, fixes #823
This fixes a small logic error. Each gateway can be multiple times in
p.recommened (for each transport). If a location was selected manually,
each gateway was added twice to the returned list. The returned list is
the list of gateways we will connect to. For more information, please
check: https://0xacab.org/leap/bitmask-vpn/-/merge_requests/213