Strip port of menshen host before resolving with logDnsLookup
In case of an error, we try to logDnsLookup(url.Host). url is based on the
provider api url. This can be localhost:8443, so we need to strip the
port :8443 before.

There is also a small change in pkg/vpn/status.go. When using an
introducer, fetching information from menshen can take some time. Then
it can be normal that the call to `GetBestLocation()` fails. This is not
a big deal, as it is called later/periodically. But when we use .Err(),
the output is printed bold to stdout and it looks like there is a big
issue.